refactor(pwa/ff): `job-data` to `job-document-storage`
