+++ uniquely/flight-finder-pwa/src/director/job-delete.ts
@@ -1,7 +1,7 @@
 import {fetch} from '@alwatr/fetch';
 import {SignalInterface} from '@alwatr/signal';
 
-import {jobDataSignal} from './job-data.js';
+import {jobDataSignal} from './job-document-storage.js';
 import {showToastSignal} from './toast.js';
 
 import type {AlwatrServiceResponse} from '@alwatr/fetch';
